+++ src/main/java/com/example/demo/service/ISurveyAnswerService.java
@@ -7,9 +7,9 @@ import java.util.List;
 
 public interface ISurveyAnswerService {
 
-    SurveyAnswerDto getById(int id);
-    SurveyAnswerDto save(Integer userId,SurveyAnswerDto surveyAnswerDto);
+    SurveyAnswerDto getById(long id);
+    SurveyAnswerDto save(long userId,SurveyAnswerDto surveyAnswerDto);
     List<SurveyAnswerDto> getAll();
-    SurveyAnswerDto update(int userId, int surveyAnswerId,SurveyAnswerDto surveyAnswerDto) throws IllegalAccessException;
-    boolean delete(int userId,int surveyAnswerId) throws IllegalAccessException;
+    SurveyAnswerDto update(long userId, long surveyAnswerId,SurveyAnswerDto surveyAnswerDto) throws IllegalAccessException;
+    boolean delete(long userId,long surveyAnswerId) throws IllegalAccessException;
 }
